configuration class added

+++ lib/rails_exception_handler.rb
@@ -9,17 +9,24 @@ class RailsExceptionHandler
     Handler.new(env, e).handle_exception
   end
 
+  def self.configuration
+    @configuration ||= Configuration.new
+  end
+
   def self.configure
-    #if Rails.env == 'production'
+    yield configuration
+    return unless configuration.environments.include?(Rails.env)
+
+    Rails.configuration.middleware.use RailsExceptionHandler
+
+    Rails.configuration.action_dispatch.show_exceptions = true
     require File.expand_path(File.dirname(__FILE__)) + '/patch/show_exceptions.rb'
+
     %w{ models controllers }.each do |dir|
       path = File.join(File.dirname(__FILE__), 'app', dir)
       $LOAD_PATH << path
       ActiveSupport::Dependencies.autoload_paths << path
       ActiveSupport::Dependencies.autoload_once_paths.delete(path)
     end
-    Rails.configuration.middleware.use RailsExceptionHandler
-    Rails.configuration.action_dispatch.show_exceptions = true
-    #end
   end
 end
